Clarifications for expression evaluation
This proposal is editorial in nature: it does not intend to change the
meaning of any DWARF constructs, but merely to clarify aspects of
DWARF expression evaluation that were unclear to the teams implementing
DWARF consumers.

The changes proposed below provide additional rigor in the description
of DWARF expressions. This proposal defines a context within which
expression evaluation takes place, specifies the conditions under
which an expression is ill formed, and adopts a more formal notation
in the description of the stack operations.