Editorial reorganization
With the changes proposed in Part 6 of this series, DWARF expressions
and location descriptions have essentially been unified, and the
distinction between the two is no longer material. The bulk of this
proposal merely reorganizes the text from Sections 2.5 (DWARF
Expressions) and 2.6 (Location Descriptions) into a single section, and
rearranges the operators into more logical groupings.