Generalize Offsetting of Location Descriptions
After [Allow location description on the DWARF evaluation stack] proposal, the
DW_OP_plus and DW_OP_minus operations can be defined to operate on a memory
location description in the default target architecture specific address space
and a generic type value to produce an updated memory location description. This
allows them to continue to be used to offset an address.

To generalize offsetting to any location description, including location
descriptions that describe when bytes are in registers, are implicit, or a
composite of these, the DW_OP_offset, DW_OP_offset_uconst, and DW_OP_bit_offset
offset operations are added.

The offset operations can operate on location storage of any size. For example,
implicit location storage could be any number of bits in size. It is simpler to
define offsets that exceed the size of the location storage as being an
evaluation error, than having to force an implementation to support potentially
infinite precision offsets to allow it to correctly track a series of positive
and negative offsets that may transiently overflow or underflow, but end up in
range. This is simple for the arithmetic operations as they are defined in
terms of two's compliment arithmetic on a base type of a fixed size. Therefore,
the offset operation define that integer overflow is ill-formed. This is in
contrast to the DW_OP_plus, DW_OP_plus_uconst, and DW_OP_minus arithmetic
operations which define that it causes wrap-around.

Having the offset operations allows DW_OP_push_object_address to push a
location description that may be in a register, or be an implicit value. The
DWARF expression of DW_TAG_ptr_to_member_type can use the offset operations
without regard to what kind of location description was pushed.

Since [Allow location description on the DWARF evaluation stack] has
generalized location storage to be bit indexable, DW_OP_bit_offset generalizes
DWARF to work with bit fields. This is generally not possible in DWARF Version
5.

The DW_OP_*piece operations only allow literal indices. A way to use a computed
offset of an arbitrary location description (such as a vector register) is
required. The offset operations provide this ability since they can be used to
compute a location description on the stack.

It could be possible to define DW_OP_plus, DW_OP_plus_uconst, and DW_OP_minus to
operate on any location description to avoid needing DW_OP_offset and
DW_OP_offset_uconst. However, this is not proposed since currently the
arithmetic operations are defined to require values of the same base type and
produces a result with the same base type. Allowing these operations to act on
location descriptions would permit the first operand to be a location
description and the second operand to be an integral value type, or vice versa,
and return a location description. This complicates the rules for implicit
conversions between default address space memory location descriptions and
generic base type values. Currently the rules would convert such a location
description to the memory address value and then perform two's compliment wrap
around arithmetic. If the result was used as a location description, it would be
implicitly converted back to a default address space memory location
description. This is different to the overflow rules on location descriptions.
To allow control, an operation that converts a memory location description to an
address integral type value would be required. Keeping a separation of location
description operations and arithmetic operations avoids this semantic
complexity.