Support for Source Languages Mapped to SIMT Hardware
If the source language is mapped onto the AMDGPU wavefronts in a SIMT manner,
then the variable DWARF location expressions must compute the location for a
single lane of the wavefront. Therefore, a DWARF operation, DW_OP_push_lane, is
added to denote the current lane, much like DW_OP_push_object_address denotes
the current object.

In addition, a way is needed for the compiler to communicate how many source
language threads of execution are mapped to a target architecture thread’s SIMT
lanes. Therefore, a DW_AT_lanes DWARF attribute is added.

> [For further discussion...]
> Combine with DWARF issue 200720.1 "SIMD location expressions".