Support for Source Languages Mapped to SIMD/SIMT Hardware
If the source language is mapped onto target architecture threads in a
SIMD and/or SIMT manner, then the variable DWARF location expressions
must compute the location for a single lane of the target architecture
threads. Therefore, a DWARF operation, DW_OP_push_lane, is added to
denote the current lane, much like DW_OP_push_object_address denotes the
current object.

In addition, a way is needed for the compiler to communicate how many
lanes are in use by the target architecture threads. Therefore, a
DW_AT_lanes DWARF attribute is added.

For example, in a vectorized SIMD execution model, the lanes of the
target architecture thread can execute multiple iterations of a loop
concurrently for a single source language thread of execution. In a SIMT
execution model, different source language threads of execution may be
mapped to separate lanes of the target architecture thread.

> [For further discussion...]
> Combine with DWARF issue 200720.1 "SIMD location expressions".