Support for Divergent Control Flow of SIMT Hardware
If the source language is mapped onto the AMDGPU wavefronts in a SIMT manner the
compiler can use the AMDGPU execution mask register to control which lanes are
active. To describe the conceptual location of non-active lanes requires an
attribute that has an expression that computes the source location PC for each
lane.

For efficiency, the expression calculates the source location the wavefront as a
whole. This can be done using the DW_OP_select_bit_piece (see [DWARF Operations
to Create Vector Composite Location Descriptions]) operation. Therefore, a
DW_AT_lane_pc DWARF attribute is added.

The AMDGPU may update the execution mask to perform whole wavefront operations.
Therefore, there is a need for an attribute that computes the current active
lane mask. This can have an expression that may evaluate to the SIMT active lane
mask register or to a saved mask when in whole wavefront execution mode.
Therefore, a DW_AT_active_lane DWARF attribute is added.